Builder Pattern Implementation for DeploymentUnit.
Currently deployment_unit is part of apps module and functions can be
performed on it which may alter the deployment scenario.
Hence deployment is moved to `deployment` module instead and
deployment_unit has a builder pattern implementation to prevent improper
access and to maintain the correct flow.
Hooks are also moved to the new deployment folder.
