+++ api/src/infrastructure/infrastructure.rs
@@ -25,6 +25,7 @@
  */
 
 use crate::config::ContainerConfig;
+use crate::deployment::DeploymentUnit;
 use crate::models::service::{Service, ServiceStatus};
 use crate::models::{ContainerType, ServiceConfig};
 use async_trait::async_trait;
@@ -84,8 +85,7 @@ pub trait Infrastructure: Send + Sync {
     async fn deploy_services(
         &self,
         status_id: &String,
-        app_name: &String,
-        strategies: &[DeploymentStrategy],
+        deployment_unit: &DeploymentUnit,
         container_config: &ContainerConfig,
     ) -> Result<Vec<Service>, Error>;
