+++ api/src/main.rs
@@ -48,6 +48,7 @@ use std::sync::Arc;
 
 mod apps;
 mod config;
+mod deployment;
 mod http_result;
 mod infrastructure;
 mod models;
